Fixed plot issue
The plots for training and testing loss used to show the same content,
because of an error in the code. For better readibility parts of the
post processing and the evalutation was moved to functions.

Results for the losses for different test train splits can now be
imported from csv files and plotted